[not verified] # This is a combination of 5 commits.
[not verified] Moved the logic to gate WPCOM blocks to a shared file

The logic to gate the blocks to paid plans on WPCOM is repeated for the
OpenTable and Calendly blocks. It is likely to be needed in future too,
so this change moves the repeated to logic to the `Jetpack_Gutenberg`
library.

Reworked incorporating ideas from #14430 and #14431

Changed availability logic so a block is always registered

Based on the suggestion by @simison, this changes the logic around block
availability, so that a block can always be registered, but its
availability can be set independently and used to prevent dynamic blocks
from rendering.

Adjusted the call to `remove_extension_prefix` because of load order

The compatibility function `wp_startswith` isn't loaded until the
`plugins_loaded` hook, but plugins sometimes call
`jetpack_register_block` before that. This caused an exception in the
`remove_extension_prefix` call.

This moves that call to the point just before the value is used,
ensuring that the function is available.

Refactored to work with the existing `Jetpack_Plan::supports` method

This is missing the WPCOM part at the moment.